+++ tsetmc/__init__.py
@@ -8,8 +8,13 @@ from typing import TypedDict as _TypedDict
 
 from aiohutils.session import SessionManager
 from jdatetime import datetime as _jdatetime
-from pandas import DataFrame as _DataFrame, read_csv as _read_csv
+from pandas import (
+    DataFrame as _DataFrame,
+    options as _o,
+    read_csv as _read_csv,
+)
 
+_o.mode.copy_on_write = True
 _csv2df = _partial(_read_csv, low_memory=False, engine='c', lineterminator=';')
 _F = r'(-?\d+(?:\.\d+)?)'  # float pattern